TIP-1073: move write value factory into the write directory
