+++ src/Akeneo/Pim/Enrichment/Component/Product/Factory/Write/Value/WriteValueFactory.php
@@ -1,0 +1,23 @@
+<?php
+declare(strict_types=1);
+
+namespace Akeneo\Pim\Enrichment\Component\Product\Factory\Read\Value;
+
+use Akeneo\Pim\Enrichment\Component\Product\Model\ValueInterface;
+use Akeneo\Pim\Structure\Component\Query\PublicApi\AttributeType\Attribute;
+
+/**
+ * The goal of this factory is create values without doing any check of the validity of the data,
+ * such as validation of a date is effectively a date or a scalar is really a scalar.
+ *
+ * It has been designed for performance purpose, as doing check on type when reading from database should not be done.
+ *
+ * @author    Anael Chardan <anael.chardan@akeneo.com>
+ * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+interface ReadValueFactory
+{
+    public function create(Attribute $attribute, ?string $channelCode, ?string $localeCode, $data): ValueInterface;
+    public function supportedAttributeType(): string;
+}
